Table column_spec style dict, config file support
Major work to support per/column styling.  Lots of interface changes
that could break backwards compat but should hopefully make things more
composable, extensible and elegant to put into use.

Initial work to support config files.  Greg wants to set his prompt to
something else.  This is always his first request when I write a CLI
suite. ;)

Along the same lines as the column styling.  Needed to improve the vt
code to support the concept of 'visual length'.  There is a lot of work
in this commit to provide a pseudo string type call VTStr that keeps
track of vt opcodes as sperate from the visual characters.  This allows
us to do column sizing for vt strings that would otherwise breakdown
when asked for their length or when consumed by a python "%s" format;
Both of which do not understand vt opcodes as being non visual
characters.