Refine data flow for Table and Tree.
Much work to get the Table and Tree layout code working well with more
diverse datasets and stream times.  Added the notion of prerendering to
the Table which does seed collection under some timing and data
constraints. The idea being it tries to get as much data as it can
within reasonable memory and timing limits.  When too much time or too
many rows have come in, it will start rendering so you don't have to
wait too long for big or slow data streams.

Support in VTML system and Table for passing in premade VTML objects and
preserveing their attributes.

Much better handling for non-terminal modes and non-ascii modes.