Working on cache tools in a data module.
Writing a time-to-live (TTL) cache mapping that is fairly specific to my
needs but generic enough that it might morph into a reusable mapping.
The mapping sits over an `OrderedDict` and expires aged-out keys
(lazily).  Because it's a cache, it has an optional max size. It's
not fancy in the least so it just expires out the oldest entry by
change time;  If you update an entry it moves to the front of the
mapping otherwise FIFO rules apply for garbage collection.

Adding a "hone" cache function decorator.  This is used to wrap tab
completer functions where the datasets generally contain super and
subsets that are 1:1 mappings of the prefix string given by a user. The
decorator will support two refinement techniques out of the box; A
'startswith' mode that roughly equates to filtering the results by
str.startswith.  It expects your datasets to be some sort of flat
content iterable (eg. {'word1', 'word2', 'otherword'}). The other mode
is `container` which is designed to traverse nested container
structures.  The expected prefix (radix) for this function is an iterable
of keys to use for traversal (eg. {"level1": {"level2": {"leaf":1}}}).
Keys for such a structure should be something like ['level1', 'level2',
'leaf'].

Like metallurgy you can take raw materials (superset) and hone them into
finer tools (subset).  The notion here is that a short radix must return the
whole superset (the most unrefined data).  If the first call is with an
empty radix then all future calls (until time-to-live is exceeded) will be
cache hits or what are called partial hits if they ran the `refineby`
routine to generate the subset.  In the case of tab completion for big
expensive requests, this lets us avoid making the same expensive call
ever again, and we benefit from that labor for the duration of the
completer session.