Add very simple HTML parser.
There are often cases where I want to print some html to the screen and
I know it won't look great, but there is a little bit of formatting I
can do to make it easier on the eyes.  So I added an html submodule to
layout that does conversion of html to vtml.  It offers similar
interfaces as the vtml module.

Added vtml.beststr() which lets you try to use unicode chars and will
failback to non unicode (or whatever your failback strings are).  Eg.

print(vtml.beststr('\u2016 try first', 'then\u900', 'and finally'))