Soften accessors def for table.
The accessors def used to take ints and strs for sequence and mapping
lookup respectively, but if the index or key was invalid it would throw
and exception.  Now it will default to an empty value.  The user can
also provide a tuple of (key, default) for the accessor definition in
cases where they want a different default value returned on lookup
failure.  The callable accessor configuration is unaffected and must
provide it's own defaulting mechanism as was previously true too.