Paging and tab completion refinement.
Move pager module to paging to avoid common name conflicts.

Support for excluding responses from command tab completion.  See
`Command.completion_excludes`.

Pass around a pager "spec" instead of a just a shell command string
for paging details.  The spec has en/dis state, pager and the tty
setting which dictates the use of vt100 escape sequences.

Moved command paging to the `Command.run_wrap` method where it is only
used on the `run` method.  This allows the user program to affect the
paging setup in `prerun` and/or do actions that are not compatible with
a pager before and after the main command code is run.