Pager cleanup.
Remove prototype debug and some finer details for paging.

Less is getting it's term size from the special LINES and COLUMNS
shell variables.  These are not updated inside the python process so we
have to hard set them in the env on each pager run to get less to
understand the proper dimensions on a terminal that changed size after
ecm was run in interactive mode.

Close stdio files on exit to catch pipe errors we wish to silence.