Add inplace and right-hand VTML concatination, misc fixes
Add proper support for inplace add of VTML strings, such as
    ```
    a1 = a2 = vtmlrender('something')
    a1 += 'more'
    assert a1 is a2
    ```
That test would fail previously because no __iadd__ was implemented and
the assignment would be a copy of the left operand.

Fix for `BrokenPipeError` in paging.  I was doing it wrong.  My test
case needed some love to better handle this exception but the command
session class was already doing the right thing for commands that use
a pager.

Fix for `VTMLBuffer.center`;  I busted it.  Also added an `rstrip=bool`
option that will optionally strip the right side of a centered output.
I intend to use this for right edge columns to avoid needless padding on
the end of a screen.