Fix for table stdout detection in some cases.
When a table is created prior to sys.stdout being changed to some other
file object, as can be the case when a pager is used, the table would
incorrectly assume it was not drawing to a tty when it came time to
print output and would not use the proper terminal size, despite using
the correct renderer class.

This patch changes the Table.file attr to be a lazy property that checks
the user override value and otherwise returns the current value of
sys.stdout.